Added Methods for analysis to CrimeData class - added average_pindexes method - added average_per_crime_pindexes method - added average_per_crime_per_neighbourhood_pindexes

+++ neighbourhood_crime.py
@@ -152,7 +152,7 @@ class NeighbourhoodCrimePIndex(NeighbourhoodCrime):
             self.p_index_dict[year] = {}
         self.p_index_dict[year][month] = p_index
 
-    def get_data(self, year: int, month: int) -> int:
+    def get_data(self, year: int, month: int) -> float:
         """Returns p-index of a given year and month
         """
         return self.p_index_dict[year][month]
